[add] material unit test [add] compiler warning switch to CMakeLists [add] access to initialized state of window
[mod] use of string_view instead of string

[fix] minor type fixes
[fix] remove-erase idiom
[fix] add EyeCandy3D library to test-project linker settings

+++ EyeCandy3D/include/EC3D/Graphics/Material/Material.h
@@ -35,8 +35,9 @@ namespace ec
 
 		/**
 		 * \brief Try to get a texture at the specified index.
+		 * \return The texture at the given index or nullptr if not existent
 		 */
-		Texture* getTexture(unsigned int index);
+		Texture* getTexture(size_t index);
 
 		/** 
 		 * \brief Get all textures associated with this material.
